+++ src/cli.ts
@@ -34,6 +34,7 @@ import { stateVerify } from "./commands/stateVerify.js";
 import { verifierCheck } from "./commands/verifierCheck.js";
 import { holdoutVerify } from "./commands/holdoutVerify.js";
 import { acceptanceVerify } from "./commands/acceptanceVerify.js";
+import { regressionCheck } from "./commands/regressionCheck.js";
 import { guidedFirstRun } from "./firstRun.js";
 
 /**
@@ -120,5 +121,6 @@ program.addCommand(stateVerify);
 program.addCommand(verifierCheck);
 program.addCommand(holdoutVerify);
 program.addCommand(acceptanceVerify);
+program.addCommand(regressionCheck);
 
 program.parseAsync(process.argv);